# Perf optimized find last non-zero element